split signalscope code into two files, partially so the signal<->frequency maps have a clearer owner that makes sense as an API elsewhere
